split flora addon into separate config and generation addons
