+++ common/addons/config-flora/src/main/java/com/dfsek/terra/addons/flora/FloraAddon.java
@@ -1,7 +1,6 @@
 package com.dfsek.terra.addons.flora;
 
 import com.dfsek.terra.addons.flora.config.BlockLayerTemplate;
-import com.dfsek.terra.addons.flora.flora.FloraLayer;
 import com.dfsek.terra.addons.flora.flora.gen.BlockLayer;
 import com.dfsek.terra.api.TerraPlugin;
 import com.dfsek.terra.api.addon.TerraAddon;
@@ -9,12 +8,9 @@ import com.dfsek.terra.api.addon.annotations.Addon;
 import com.dfsek.terra.api.addon.annotations.Author;
 import com.dfsek.terra.api.addon.annotations.Version;
 import com.dfsek.terra.api.event.EventListener;
-import com.dfsek.terra.api.event.events.config.ConfigurationLoadEvent;
 import com.dfsek.terra.api.event.events.config.pack.ConfigPackPreLoadEvent;
 import com.dfsek.terra.api.injection.annotations.Inject;
 import com.dfsek.terra.api.registry.exception.DuplicateEntryException;
-import com.dfsek.terra.api.util.seeded.BiomeBuilder;
-import com.dfsek.terra.api.world.generator.GenerationStageProvider;
 
 @Addon("config-flora")
 @Author("Terra")
@@ -30,14 +26,6 @@ public class FloraAddon extends TerraAddon implements EventListener {
 
     public void onPackLoad(ConfigPackPreLoadEvent event) throws DuplicateEntryException {
         event.getPack().registerConfigType(new FloraConfigType(event.getPack()), "FLORA", 2);
-        event.getPack().getOrCreateRegistry(GenerationStageProvider.class).register("FLORA", pack -> new FloraPopulator(main));
-        event.getPack().applyLoader(FloraLayer.class, FloraLayerLoader::new)
-                .applyLoader(BlockLayer.class, BlockLayerTemplate::new);
-    }
-
-    public void onBiomeLoad(ConfigurationLoadEvent event) {
-        if(BiomeBuilder.class.isAssignableFrom(event.getType().getTypeClass())) {
-            event.getLoadedObject(BiomeBuilder.class).getContext().put(event.load(new BiomeFloraTemplate()).get());
-        }
+        event.getPack().applyLoader(BlockLayer.class, BlockLayerTemplate::new);
     }
 }
