+++ common/addons/config-flora/src/main/java/com/dfsek/terra/addons/flora/FloraFactory.java
@@ -1,6 +1,6 @@
 package com.dfsek.terra.addons.flora;
 
-import com.dfsek.terra.addons.flora.flora.TerraFlora;
+import com.dfsek.terra.addons.flora.flora.gen.TerraFlora;
 import com.dfsek.terra.api.TerraPlugin;
 import com.dfsek.terra.api.config.ConfigFactory;
 import com.dfsek.terra.api.world.Flora;
