+++ common/addons/config-flora/src/main/java/com/dfsek/terra/addons/flora/FloraTemplate.java
@@ -3,8 +3,8 @@ package com.dfsek.terra.addons.flora;
 import com.dfsek.tectonic.annotations.Default;
 import com.dfsek.tectonic.annotations.Final;
 import com.dfsek.tectonic.annotations.Value;
-import com.dfsek.terra.addons.flora.flora.TerraFlora;
 import com.dfsek.terra.addons.flora.flora.gen.BlockLayer;
+import com.dfsek.terra.addons.flora.flora.gen.TerraFlora;
 import com.dfsek.terra.api.config.AbstractableTemplate;
 import com.dfsek.terra.api.util.collection.MaterialSet;
 import com.dfsek.terra.api.util.seeded.NoiseSeeded;
