+++ common/addons/generation-stage-flora/src/main/java/com/dfsek/terra/addons/generation/flora/BiomeFloraTemplate.java
@@ -1,10 +1,8 @@
-package com.dfsek.terra.addons.flora;
+package com.dfsek.terra.addons.generation.flora;
 
 import com.dfsek.tectonic.annotations.Default;
 import com.dfsek.tectonic.annotations.Value;
-import com.dfsek.tectonic.config.ConfigTemplate;
 import com.dfsek.tectonic.loading.object.ObjectTemplate;
-import com.dfsek.terra.addons.flora.flora.FloraLayer;
 
 import java.util.Collections;
 import java.util.List;
