+++ common/addons/generation-stage-flora/src/main/java/com/dfsek/terra/addons/generation/flora/FloraGenerationAddon.java
@@ -6,7 +6,11 @@ import com.dfsek.terra.api.addon.annotations.Addon;
 import com.dfsek.terra.api.addon.annotations.Author;
 import com.dfsek.terra.api.addon.annotations.Version;
 import com.dfsek.terra.api.event.EventListener;
+import com.dfsek.terra.api.event.events.config.ConfigurationLoadEvent;
+import com.dfsek.terra.api.event.events.config.pack.ConfigPackPreLoadEvent;
 import com.dfsek.terra.api.injection.annotations.Inject;
+import com.dfsek.terra.api.util.seeded.BiomeBuilder;
+import com.dfsek.terra.api.world.generator.GenerationStageProvider;
 
 @Addon("generation-stage-flora")
 @Version("1.0.0")
@@ -19,5 +23,16 @@ public class FloraGenerationAddon extends TerraAddon implements EventListener {
     @Override
     public void initialize() {
         main.getEventManager().registerListener(this, this);
+    }
+
+    public void onPackLoad(ConfigPackPreLoadEvent event) {
+        event.getPack().applyLoader(FloraLayer.class, FloraLayerLoader::new);
+        event.getPack().getOrCreateRegistry(GenerationStageProvider.class).register("FLORA", pack -> new FloraPopulator(main));
+    }
+
+    public void onBiomeLoad(ConfigurationLoadEvent event) {
+        if(BiomeBuilder.class.isAssignableFrom(event.getType().getTypeClass())) {
+            event.getLoadedObject(BiomeBuilder.class).getContext().put(event.load(new BiomeFloraTemplate()).get());
+        }
     }
 }
