+++ common/addons/generation-stage-flora/src/main/java/com/dfsek/terra/addons/generation/flora/FloraPopulator.java
@@ -1,6 +1,5 @@
-package com.dfsek.terra.addons.flora;
+package com.dfsek.terra.addons.generation.flora;
 
-import com.dfsek.terra.addons.flora.flora.FloraLayer;
 import com.dfsek.terra.api.TerraPlugin;
 import com.dfsek.terra.api.profiler.ProfileFrame;
 import com.dfsek.terra.api.util.PopulationUtil;
@@ -8,7 +7,6 @@ import com.dfsek.terra.api.vector.Vector2;
 import com.dfsek.terra.api.world.Chunk;
 import com.dfsek.terra.api.world.TerraWorld;
 import com.dfsek.terra.api.world.World;
-import com.dfsek.terra.api.world.biome.TerraBiome;
 import com.dfsek.terra.api.world.biome.generation.BiomeProvider;
 import com.dfsek.terra.api.world.generator.TerraGenerationStage;
 import org.jetbrains.annotations.NotNull;
